refactor: all file types now use kebab-case; removed barrel file for imports; updated imports everywhere to handle the updates
