+++ src/app/layout.tsx
@@ -3,10 +3,10 @@ import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import Header from '@/components/header';
 import Footer from '@/components/footer';
-import { LazyMotionWrapper } from '@/components/ui/LazyMotionWrapper';
+import { LazyMotionWrapper } from '@/components/ui/lazy-motion-wrapper';
 import { PostHogProvider } from './providers';
-import { PageTracker } from '../components/PageTracker';
-import { AnalyticsConsentBanner } from '@/components/ui/AnalyticsConsentBanner';
+import { PageTracker } from '../components/page-tracker';
+import { AnalyticsConsentBanner } from '@/components/ui/analytics-consent-banner';
 
 const inter = Inter({ subsets: ['latin'] });
