+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import HeroSection from '@/components/hero-section';
 import InsightsSection from '@/components/insights-section';
 import FeaturesSection from '@/components/features-section';
 import ContactSection from '@/components/contact-section';
-import AnimatedSection from '@/components/ui/AnimatedSection';
+import AnimatedSection from '@/components/ui/animated-section';
 
 export default function Home() {
   return (
