--- src/components/case-study/index.ts
@@ -1,5 +1,0 @@
-export { default as CaseStudyHero } from './case-study-hero';
-export { default as CaseStudyOverview } from './case-study-overview';
-export { default as CaseStudySection } from './case-study-section';
-export { default as CaseStudyImage } from './case-study-image';
-export { default as CaseStudyQuote } from './case-study-quote';
