--- src/components/contact/index.ts
@@ -1,7 +1,0 @@
-export { default as ContactForm } from './contact-form';
-export { default as ContactHero } from './contact-hero';
-export { default as ContactInfo } from './contact-info';
-export { default as FaqItem } from './faq-item';
-export { default as FaqSection } from './faq-section';
-export { default as FaqPreview } from './faq-preview';
-export { default as TrackedFAQLink } from './TrackedFAQLink';
