+++ src/components/footer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import SocialButtons from './footer/SocialButtons';
-import TrackedNavigation from './footer/TrackedNavigation';
-import LegalLinks from './footer/LegalLinks';
+import SocialButtons from './footer/social-buttons';
+import TrackedNavigation from './footer/tracked-navigation';
+import LegalLinks from './footer/legal-links';
 
 export default function Footer() {
   const companyNavigation = [
