--- src/components/footer/index.ts
@@ -1,3 +1,0 @@
-export { default as SocialButtons } from './SocialButtons';
-export { default as TrackedNavigation } from './TrackedNavigation';
-export { default as LegalLinks } from './LegalLinks';
