+++ src/components/footer/social-buttons.tsx
@@ -2,8 +2,8 @@
 
 import React from 'react';
 import { FaLinkedin, FaGithub } from 'react-icons/fa';
-import { IconButton } from '@/components/ui';
-import { useEventTracking } from '@/lib/hooks/useEventTracking';
+import { IconButton } from '@/components/ui/icon-button';
+import { useEventTracking } from '@/lib/hooks/use-event-tracking';
 
 export default function SocialButtons() {
   const { trackNavigation } = useEventTracking();
