--- src/components/insights/index.ts
@@ -1,1 +1,0 @@
-export { default as InsightCard } from './insight-card';
