+++ src/components/ui/animated-section.tsx
