--- src/components/ui/index.ts
@@ -1,7 +1,0 @@
-export { Button, buttonVariants } from './button';
-export { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from './card';
-export { Input } from './input';
-export { Textarea } from './textarea';
-export { IconButton, iconButtonVariants } from './icon-button';
-export { default as AnimatedSection } from './AnimatedSection';
-export { AnalyticsConsentBanner } from './AnalyticsConsentBanner';
