--- src/lib/hooks/index.ts
@@ -1,7 +1,0 @@
-export { useClientConfig } from './useClientConfig';
-export { useRecaptcha } from './useRecaptcha';
-export { useScrollPosition } from './useScrollPosition';
-export { usePageTracking } from './usePageTracking';
-export { useEventTracking } from './useEventTracking';
-export { useUserIdentification } from './useUserIdentification';
-export { useAnalyticsConsent } from './useAnalyticsConsent';
