+++ src/lib/hooks/use-analytics-consent.ts
