+++ src/lib/hooks/use-page-tracking.ts
